chore: move ConfigureOCR3ContractFromJD
